re-bundle Test::More from 1.00002 to fix test errors on Perl 5.10.0
The bundled Test::More code was initially removed to reduce
redundant and unmaintained code in favor of the officially supported
and maintained version on CPAN.

However, the Test::More distribution available from CPAN circa the
year 2025 causes a test failure on Perl version 5.10.0 (but not other
versions) for tied $@.

5.10.0 is long unsupported and has a number of bugs fixed in 5.10.1.
It's not worth identifying the issues which cause the CPAN Test::More
distribution to act differently on 5.10.0.

It would be inappropriate to ship a new version of Sub::Delete
which fails on a previously supported Perl version because of
issues with a test framework, when the original test framework
suffices to test the code.

With that in mind, and in the spirit of making minimal changes to this
distribution to achieve the primary goal of remedying the deprecation
of the apostrophe as a module separator, the original bundled
Test::More is restored.